Add clever Artist extraction from song title
Artist is parsed from the song filename by following regex: (.+) - (.+).
It works for titles like 'Aerosmith - I Don't Want To Miss A Thing.mp3' but
doesn't work for '09 - Lappi - I. Eramaajarvi.mp3'.
From now on we'll determine whether major part of album songs has artist in the
title. If not then we clear Artist in all songs that have it.