Disable changing of track number when multiple songs are edited
Previously, special text was used in text box for track number - "<blank>" or "<keep>".
If user was changing the value, he had to erase this text and put some number. This looked weird.
The better approach would be to print this special text as placeholder, which is automatically cleared when user clicks the text box.
However, seems like the edit of track number in multi-song mode is unlikely.
That is why we disable changing of track number when multiple songs are edited, like it's done for song title.